+++ client/app/account/user/security/u2f/user-security-u2f.controller.js
@@ -1,4 +1,4 @@
-angular.module("UserAccount.controllers").controller("UserAccount.controllers.doubleAuth.u2f", [
+angular.module("UserAccount").controller("UserAccount.controllers.doubleAuth.u2f", [
     "$scope",
     "$q",
     "$translate",