+++ client/app/account/user/ssh/view/user-ssh-view.controller.js
@@ -1,3 +1,3 @@
-angular.module("UserAccount.controllers").controller("UserAccount.controllers.ssh.view", ($scope) => {
+angular.module("UserAccount").controller("UserAccount.controllers.ssh.view", ($scope) => {
     $scope.data = $scope.currentActionData;
 });