+++ client/app/cdn/dedicated/manage/ssl/generate/cdn-dedicated-ssl-generate.controller.js
@@ -1,31 +1,82 @@
-angular.module("App").controller("CdnGenerateSslCtrl", ($scope, $stateParams, $translate, Cdn, Alerter) => {
-    "use strict";
+angular.module("App").controller("CdnGenerateSslCtrl", class CdnGenerateSslCtrl {
 
-    $scope.ssl = null;
+    constructor ($state, $stateParams, $translate, OvhApiCdn, Alerter) {
+        // dependencies injections
+        this.$state = $state;
+        this.$stateParams = $stateParams;
+        this.$translate = $translate;
+        this.OvhApiCdn = OvhApiCdn;
+        this.Alerter = Alerter;
 
-    $scope.loadSsl = function () {
-        $scope.loading = true;
-        Cdn.getSsl($stateParams.productId)
-            .then((ssl) => {
-                $scope.ssl = ssl.status === null ? null : ssl;
-            })
-            .catch((error) => {
-                error.message = error.message.replace(" : null", "");
-                $scope.setMessage($translate.instant("cdn_configuration_add_ssl_get_error"), { type: "ERROR", message: error.message });
-            })
-            .finally(() => {
-                $scope.loading = false;
+        // controller attributes
+        this.model = {
+            name: null
+        };
+
+        this.loading = {
+            init: false,
+            generate: false
+        };
+
+        this.ssl = null;
+        this.actionEnabled = true;
+    }
+
+    /* =============================
+    =            EVENTS            =
+    ============================== */
+
+    onCdnSslGenerateFormSubmit () {
+        if (this.cdnSslGenerateForm.$invalid) {
+            return false;
+        } else if (!this.actionEnabled) {
+            return this.$state.go("^");
+        }
+
+        this.loading.generate = true;
+
+        return this.OvhApiCdn.Dedicated().Ssl().v6().save({
+            serviceName: this.$stateParams.productId
+        }, this.model).$promise.then(() => {
+            this.Alerter.success(this.$translate.instant("cdn_dedicated_ssl_generate_success"), "cdnDedicatedManage");
+            this.$state.go("^", {}, {
+                reload: true
             });
-    };
+        }).catch((error) => {
+            this.Alerter.error([this.$translate.instant("cdn_dedicated_ssl_generate_error"), _.get(error, "data.message")].join(" "), "cdnDedicatedManage");
+            this.$state.go("^");
+        }).finally(() => {
+            this.loading.generate = false;
+        });
+    }
 
-    $scope.entry = {};
+    /* -----  End of EVENTS  ------ */
 
-    $scope.addSsl = function () {
-        $scope.resetAction();
-        Cdn.addSsl($stateParams.productId, $scope.entry)
-            .then(() => Alerter.success($translate.instant("cdn_configuration_generate_ssl_success"), "cdn_dedicated"))
-            .catch((err) => Alerter.alertFromSWS($translate.instant("cdn_configuration_generate_ssl_fail"), err, "cdn_dedicated"));
-    };
+    /* =====================================
+    =            INITIALIZATION            =
+    ====================================== */
 
-    $scope.loadSsl();
+    $onInit () {
+        this.loading.init = true;
+
+        return this.OvhApiCdn.Dedicated().Ssl().v6().get({
+            serviceName: this.$stateParams.productId
+        }).$promise.then((ssl) => {
+            this.ssl = ssl;
+            this.model.name = this.ssl.name;
+            this.actionEnabled = this.ssl.status === "on" || this.ssl.status === "off";
+        }).catch((error) => {
+            if (_.get(error, "status") === 404) {
+                this.actionEnabled = true;
+                return null;
+            }
+
+            return this.Alerter.error([this.$translate.instant("cdn_dedicated_ssl_generate_load_error"), _.get(error, "data.message")].join(" "), "cdnDedicatedManage");
+        }).finally(() => {
+            this.loading.init = false;
+        });
+    }
+
+    /* -----  End of INITIALIZATION  ------ */
+
 });