+++ client/app/components/feature-availability/feature-availability.service.js
@@ -147,6 +147,14 @@ class FeatureAvailability {
         return this.deny("US");
     }
 
+    hasVrackAccessibleFromSidebar () {
+        return this.allow("US");
+    }
+
+    allowVrackOrder () {
+        return this.allow("US");
+    }
+
     allow (...args) {
         return Array.from(args).indexOf(this.target) > -1;
     }