+++ client/app/components/sidebar-menu/sidebar-menu.config.js
@@ -1,8 +1,11 @@
 angular.module("App")
-    .run(($q, SidebarMenu, Products, User, $translatePartialLoader, $translate, DedicatedCloud, Nas, CdnDomain, featureAvailability) => {
+    .run(($q, SidebarMenu, Products, User, $translatePartialLoader, $translate, DedicatedCloud, Nas, CdnDomain, featureAvailability, constants) => {
 
         function buildSidebarActions () {
-            return User.getUrlOf("dedicatedOrder").then((dedicatedOrderUrl) => {
+            return $q.all({
+                dedicatedOrder: User.getUrlOf("dedicatedOrder"),
+                vrackOrder: User.getUrlOf("vrackOrder")
+            }).then((results) => {
                 const actionsMenuOptions = [];
 
                 if (featureAvailability.hasNas()) {
@@ -22,9 +25,18 @@ angular.module("App")
                 actionsMenuOptions.push({
                     title: $translate.instant("navigation_left_dedicatedServers"),
                     icon: "ovh-font ovh-font-server",
-                    href: dedicatedOrderUrl,
+                    href: results.dedicatedOrder,
                     target: "_blank"
                 });
+
+                if (featureAvailability.allowVrackOrder()) {
+                    actionsMenuOptions.push({
+                        title: $translate.instant("navigation_left_vrack"),
+                        icon: "ovh-font ovh-font-vRack",
+                        href: results.vrackOrder,
+                        target: "_blank"
+                    });
+                }
 
                 return SidebarMenu.addActionsMenuOptions(actionsMenuOptions);
             });
@@ -98,6 +110,16 @@ angular.module("App")
                 state: "app.ip",
                 icon: "ovh-font ovh-font-ip"
             });
+
+            if (featureAvailability.hasVrackAccessibleFromSidebar()) {
+                SidebarMenu.addMenuItem({
+                    name: "vrack",
+                    title: $translate.instant("navigation_left_vrack"),
+                    url: constants.vrackUrl,
+                    target: "_self",
+                    icon: "ovh-font ovh-font-vRack"
+                });
+            }
 
             const productsPromise = Products.getProductsByType().then((products) => {
                 const pending = [];