+++ client/app/dedicatedCloud/dedicatedCloud.routes.js
@@ -1,6 +1,6 @@
-angular.module("App").config(($stateProvider) => {
+angular.module("App").config(($stateProvider, $urlServiceProvider) => {
     $stateProvider.state("app.dedicatedClouds", {
-        url: "/configuration/dedicated_cloud/:productId?action&token",
+        url: "/configuration/dedicated_cloud/:productId",
         views: {
             "": {
                 templateUrl: "dedicatedCloud/dedicatedCloud.html",
@@ -12,5 +12,15 @@ angular.module("App").config(($stateProvider) => {
         },
         reloadOnSearch: false,
         translations: ["dedicatedCloud"]
+    });
+
+    // ensure compatibility with links sended by emails like #/configuration/dedicated_cloud/pcc-123456?action=confirmcancel&token=myToken
+    // make a redirect to the new url of ui route
+    $urlServiceProvider.rules.when("/configuration/dedicated_cloud/:productId?action&token", (match) => {
+        if (match.action === "confirmcancel") {
+            return `/configuration/dedicated_cloud/${match.productId}/terminate-confirm?token=${match.token}`;
+        }
+
+        return false;
     });
 });